// frontend/src/app/api/model/reunion.ts
/**
 * API oPROsita
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


export interface Reunion { 
    id?: number;
    titulo?: string;
    descripcion?: string;
    fechaHora?: string;
    enlace?: string;
    grupoId?: number;
}